backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.16.0
alembic==1.13.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
markdown==3.5.1
pypandoc==1.12

# AI/ML
langchain==0.1.0
langchain-openai==0.0.2
openai==1.7.1
tiktoken==0.5.2
chromadb==0.4.22
weaviate-client==3.26.0

# NLP
spacy==3.7.2
nltk==3.8.1

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0